9.109. JPA Delete by Id Not Found

+++ src/main/java/guru/springframework/spring7restmvc/services/BeerServiceJpa.java
@@ -66,9 +66,14 @@ public class BeerServiceJpa implements BeerService {
     }
 
     @Override
-    public void deleteById(UUID beerId) {
+    public Boolean deleteById(UUID beerId) {
 
-        beerRepository.deleteById(beerId);
+        if (beerRepository.existsById(beerId)) {
+            beerRepository.deleteById(beerId);
+            return true;
+        }
+
+        return false;
     }
 
     @Override
